Revert ""Magic overhead" changed from 8 to 28"
This reverts commit 6d120f8c28b8c0e41e20f054ef19c0b754aa2004.

Adding 28 depends on the formula as written for SX1276, which is written in this
confusing way:

`+ 28 - 20IH` where (IH = 1 for explicit header, 0 otherwise)

For SX1262 it is written differently and we have the equivalent (and clearer):

`+ 8 + Nsymbolheader` where Nsymbolheader = 20 for explicit header and 0 otherwise.

Therefore the original calculation is right, because 20 is added again later for
the header.